refactor: also re-use context
